+++ app/models/follower.rb
@@ -2,6 +2,6 @@
 
 # Class for the follower db entry
 class Follower < ApplicationRecord
-  belongs_to :follower_user, :class_name => 'User'
-  belongs_to :followed_user, :class_name => 'User'
+  belongs_to :user, :class_name => 'User'
+  belongs_to :following, :class_name => 'User'
 end
