+++ app/views/users/_editform.html.slim
@@ -93,11 +93,11 @@
             - @followers.each do |f|
               div.following-table-cell
                 div.followed-user-cell
-                  a href="/users/#{f.followed_user.id}"
-                    = image_tag f.followed_user.profile_picture
-                  a href="/users/#{f.followed_user.id}"
-                    p #{f.followed_user.username} 
-                = render partial: "application/follower_button", locals: {user: f.followed_user}
+                  a href="/users/#{f.following.id}"
+                    = image_tag f.following.profile_picture
+                  a href="/users/#{f.following.id}"
+                    p #{f.following.username} 
+                = render partial: "application/follower_button", locals: {user: f.following}
           - else
             b You aren't following anyone
