+++ db/migrate/20201123051326_create_followers.rb
@@ -1,0 +1,10 @@
+class CreateFollowers < ActiveRecord::Migration[6.0]
+  def change
+    create_table :followers do |t|
+      t.references :user
+      t.references :following
+
+      t.timestamps
+    end
+  end
+end
