+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_10_03_223128) do
+ActiveRecord::Schema.define(version: 2020_11_23_051326) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -84,12 +84,12 @@ ActiveRecord::Schema.define(version: 2020_10_03_223128) do
   end
 
   create_table "followers", force: :cascade do |t|
-    t.bigint "follower_user_id"
-    t.bigint "followed_user_id"
+    t.bigint "user_id"
+    t.bigint "following_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.index ["followed_user_id"], name: "index_followers_on_followed_user_id"
-    t.index ["follower_user_id"], name: "index_followers_on_follower_user_id"
+    t.index ["following_id"], name: "index_followers_on_following_id"
+    t.index ["user_id"], name: "index_followers_on_user_id"
   end
 
   create_table "house_participations", force: :cascade do |t|
@@ -254,8 +254,6 @@ ActiveRecord::Schema.define(version: 2020_10_03_223128) do
     t.index ["name"], name: "index_users_on_name", unique: true
   end
 
-  add_foreign_key "followers", "users", column: "followed_user_id"
-  add_foreign_key "followers", "users", column: "follower_user_id"
   add_foreign_key "house_participations", "users"
   add_foreign_key "moderator_applications", "users"
   add_foreign_key "moderator_logs", "users"
